Allow compilation on Linux
For compilation on Linux some minor changes have to be made, mostly due
to the new use of `qwt-qt5`, and the use of a different compiler. On
Linux, we can use the normal `qwt` and `qwt-qt5` is not available. Linux
uses dynamic compilation of the phast executable, without a
libphast_gui.dll.a equivalent. And on Linux a function named "`index`"
is defined in `strings.h`, so we cannot create a type with the same
name. This was changed to `index_t`. Qutools does not provide Linux
libraries for the quTAU and quTAG standard because our devices are
outdated. We must therefore remove their support.